rules: new macros ptx/config-version and ptx/config-md5
There are several packages that get the version/md5 from the ptxconfig
variables PTXCONF_<PKG>_VERSION / PTXCONF_<PKG>_MD5. These new new macros
can be used for this and add basic sanity checking:

If the package is enable, then the version and md5 must not be empty.
Otherwise the string 'undefined' is returned to make it obvious that the
version is currently not available.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>