rules: add version to some dummy packages
All versions should have a version. This will be needed to find broken
packages with lint:
If a target package is removed, then the corresponding host package may be
forgotton. It won't have a version any more so it can be found this way.

Adding versions here for dummy packages will avoid false positives for this
check.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>