ptxd_make_xpkg_pkg: limit install_alternative_tree permissions
install_alternative_tree takes the permissions from the filesystem. This
can be a problem because the permission may change depending on the build
machine configuration:
The files are most likely created by some version control software. And in
this case the permissions depend on the current umask.
The most common umasks are 0022 and 0002. This results in directory
permissions 0755 and 0775.
To make this more deterministic, the permssions for
install_alternative_tree are limited by an 0022 umask. In the example
above, this results in 0755 directory permissions in both cases.

Note: This does not help in all cases: If the files and directories are
created with a even more restrictive umask then the resulting permissions
will be different. But that is less common. So this fix should help in most
cases.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>