ptxd_kconfig_update: add a check when looking for Kconfig
For legacy reasons, we allow Kconfig top-level in the BSP.
If the top layer has config Kconfig and a lower layer has a top-level
Kconfig then the wrong file (from the lower layer) is used.

Fix this by dropping support for the legacy location and fail with an error
is a file is found in that location.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>