patches/autogen.sh: fake autopoint if necessary
Autopoint just copies ome files. If the archive was created with
'make dist' then those file already exist. So autopoint is not actually
needed. So if host-gettext (which provides autopoint) is not in the
dependency list then fake it by setting AUTOPOINT=true.
This avoids a dependency on host-gettext just for autoreconf.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>