ptxdist: reorder icecc environment creation
Without this, clang may be missing from the environment.

The problem is, that the real/clang symlink is provided by
ptxd_lib_setup_target_wrapper because the clang from the toolchain can be
used as a host compiler as well.
So run ptxd_lib_setup_host_icecc later to ensure that this symlink was
already created.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>