dtc: add package to EXTRA_PACKAGES on !PTXCONF_DTC_INSTALL_OFTREE
If the oftree is not installed into the root filesystem, we don't need
to add it to PACKAGES. Add it to EXTRA_PACKAGES in this case instead.

This is useful for oftrees in FIT images (KERNEL_FIT). It allows adding
custom image dependencies for the dtc package for rather special use
cases.

Signed-off-by: Bastian Krause <bst@pengutronix.de>
Message-Id: <20210407160114.29806-1-bst@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>