bind: new package
Newer ISC DHCP versions try to build a packaged BIND library, but fail
at it. Outsource the BIND libs into an extra recipe so we can set the
configure options and BUILD_CC appropriately. This package only builds
the static BIND libraries for now, which is all that is needed by the
DHCP package.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>

Message-Id: <20210524183500.32170-2-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>