dhcp: version bump 4.1.2 -> 4.4.2
Don't use the in-tree bind because dhcp's make will try to configure it
in the compile stage. Zlib is now required too. Clean up the old-style
variable names too, and pin down all configure options. The package now
builds fine without the need for removing the -Werror compiler option.

The license was changed to MPL-2.0 in most places, also add the
respective license MD5 sums.

A patch is needed to make it build with GCC 10.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20210524183500.32170-3-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>