minicom: version bump 2.6.2. -> 2.8
Add two patches for reproducible builds and compatibility with GCC 11,
renovate the variable names, sort the config options in the order in
which they appear in configure.ac for configure_helper.py, and pin them
down so that configure does not try to autodetect paths on the build
host. Also make it work with ckermit if desired.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20210524212129.8277-2-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>