rules: rework some error handling
There are several places where 'make' is aborted with an '$(error ...)'
immediately when starting ptxdist. It's useful to detect problems early on.
However, this means, that it also interferes with 'ptxdist print' to it can
make debugging issues a bit more tricky.

So collect the errors instead and print them in a early make target
instead.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>