rework device-tree handling
It's not really possible to build a dtb by itself. A lot of included dtsi
and header files are needed. So build the device-trees with the package
that provides these files. Currently that's the kernel package.

This also means, that other kernel packages can have their own device-trees
as well.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>