shaderc: drop patch
This is actually caused by a gcc bug[1]. No OSELAS.Toolchain is affected by
this. And the only default host compiler that has this issue is on Arch.
And several other packages fail to built because of this as well. So lets
just drop the patch. The issue will go away soon anyways.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100392

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>