fscryptctl: new package
fscryptctl is a low-level tool that handles raw keys and manages
policies for the Linux filesystem encryption, specifically the
"fscrypt" kernel interface which is supported by the ext4,
f2fs, and UBIFS filesystems.

fscryptctl is mainly intended for embedded systems which can't use the
full-featured fscrypt tool.  It does *not* handle key generation,
key stretching, key wrapping, or PAM integration.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Message-Id: <20210802054231.6401-1-a.fatoum@pengutronix.de>
[mol: expand license text for fscryptctl.c to start earlier]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>