ptxd_make_world_update_md5: handle layers
When layers are involved, then the delta config must be updated as well.
Either update the existing symbol or add a new one at the end of the file.

If it is added at the end, then print a warning, that 'oldconfig' should be
used to correct the symbol oder in the file.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>