openssh: rc-once: Do not overwrite existing keys
When storing your keys not in rootfs but on a separate data partition
(using symbolic links or overlay fs), keys are overwritten on each
firmware upgrade which lets rc-once run again (which happens when using
opkg upgrade/update or RAUC in an A/B scheme for example).

Changing keys are at best annoying, but may be interpreted as an attack
as well.

Note: if an empty or invalid key is found, the key is still regenerated.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Message-Id: <20211105154734.19983-9-ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>