kernel: avoid config changes based on the availability of pahole
CONFIG_PAHOLE_HAS_SPLIT_BTF shows up in the config only when pahole is
installed in the system. As a result, the config changes depending on
whether pahole is installed or not.
Set PAHOLE=false to ensure that it is never found.

If this is actually needed in the future, then we can add an option for it
and create a host package for pahole.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>