dropbear: license review
Commit 80326bc43e7 (2020-11-16, Alexander Dahl: "dropbear: version bump
2019.78 -> 2020.79") mentioned:

    The curve25519 implementation was replaced with the one by TweetNaCl,
    which induced a change in 'LICENSE' summary.  The old implementation was
    licensed BSD-3-Clause and the new is public domain.

This was wrong; even though curve25519.c is based on the public domain
TweetNaCl library, as mentioned in LICENSE, it carries an MIT
license header like most other files. Therefore remove the first
public_domain sub-expression.

The second public_domain sub-expression referred to the libtomcrypt
license, which is licensed as, alternatively, WTFPL, or "LibTomCrypt is
public domain.  As should all quality software be.". Classify it as a
custom license, as we deprecated the public_domain license identifier in
commit 76d1f680233955839298 (2021-08-06, Roland Hieber: "doc: working
with licensing information in packages").

Fixes: 80326bc43e71d1b012b3 ("dropbear: version bump 2019.78 -> 2020.79")
Cc: Alexander Dahl <ada@thorsis.com>
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20211205221106.240253-1-rhi@pengutronix.de>
Acked-by: Alexander Dahl <ada@thorsis.com>
[mol: fixed commit message as suggested by Alexander Dahl]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>