gt: version bump 31b13554... -> 2021-09-30-g7247547a
Update to the latest master to get a few build system cleanups. For
example, the install location of the config file is now /etc, which
allows us to use install_alternative to be able to override the config
in a BSP.

Also add the date to the package version and document that there are no
tags for gt.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20220106105934.1598955-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>