barebox: add integration of firmware blobs
In some cases barebox requires firmware blobs, which may be
provided in binary form by the vendor or compiled in a
preceding step. Add the possibility to specify files (in
separate rule files) which are injected in the barebox source
directory during preparation.

The virtual symbol BAREBOX_DEPENDENCIES is introduced in this
patch to allow for the addition of dependencies on e.g.,
firmware packages.

Signed-off-by: Michael Riesch <michael.riesch@wolfvision.net>
Message-Id: <20220129070330.2601433-5-michael.riesch@wolfvision.net>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>