linux-firmware: Add new package. Version 20211216
Firmware is needed by a lot of modern devices.
Includes various radio interfaces, GPU and CPU stuff.

You need to manually select license files that apply
to the firmware files you are copying.
This is due to the nature of the firmware package.
There is no strict relation between firmware file name
and license file name.

This is more generic than having to maintain a configuration
set for every type of firmware file or function.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20220131105449.3188353-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>