weston: version bump 9.0.0 -> 10.0.0
Drop the WESTON_DRM_COMPOSITOR option and always build the drm backend.
It was unconditionally selected as the default anyway and not building
the backend results in a build failure.

Drop WESTON_LAUNCH, as the weston-launch is deprecated and was BROKEN
anyway.

Add WESTON_SEATD to enable the libseat-launcher.

Add WESTON_DEPRECATED_WL_SHELL, which defaults to false to enable the
deprecated wl-shell if clients still require it.

Remove BROKEN from WESTON_PIPEWIRE, as PTXdist now includes Pipewire.

Update the installation of the example config files to use the files
that are generated by Weston instead of generating it in the rule file

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20220207114645.3916882-1-m.tretter@pengutronix.de>
[mol: remove last reference to PTXCONF_WESTON_LAUNCH]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>