tf-a: Support enabling DEBUG via TF_A_EXTRA_ARGS
With DEBUG=1 the resulting images are located in a different folder, so
adapt the install target accordingly. Before this adaption the build
either failed when DEBUG=1 was contained in TF_A_EXTRA_ARGS (because no
images were found in the release folder) or (IMHO worse) ptxdist picked
up the build results of an earlier attempt where DEBUG
wasn't set yet in TF_A_EXTRA_ARGS.

Also remove DEBUG=0 from the hardcoded set of options. This is the
default in the tf-a Makefiles anyhow, and passing both DEBUG=0 and
DEBUG=1 only results in confusion.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Message-Id: <20220309111852.537125-1-u.kleine-koenig@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>