libxml2: version bump 2.9.12 -> 2.9.13
- Security fixes for CVE-2022-23308, use-after-free, null-deref, and memory leaks
- Regression fixes
- Bug fixes
- Improvements in code quality
- Build system fixes
- Documentation updates

One upstream patch had to be added for fixing a build error if
PTXCONF_LIBXML2_DTD is not set and --without-valid is passed to
configure.

The download location and tarball compression format changed.
While the old location is still there, it does not have the new tarball.
The new location however has only tarballs in .xz format, no .gz
anymore.  Thus both have to be changed.

Link: https://mail.gnome.org/archives/xml/2022-February/msg00009.html
Signed-off-by: Alexander Dahl <ada@thorsis.com>
Message-Id: <20220309163207.36477-1-ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>