u-boot: Add target name options.
Add possibility to remove install of the normal u-boot.bin.
A lot of targets do not need it, some targets do not even build it
(Which generates an error during the install).
Add a i.mx8 family flash.bin image name.

This entire thing bugs me a bit.
I don't really like the idea of a lot of target types just to copy files.
Usually I know what the target file is named and what I want it installed as
(which sometimes is a more generic name than the build name).
I can see this disappearing in favor of a more generic "copy this" "to this" with
the default names filled in and a help listing the typical binary target names
as a guideline.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20220310115612.1404060-1-christian.melki@t2data.com>
[mol: 'default y' instead of 'default true']
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>