libcap-ng: Version bump. 0.8.2 -> 0.8.3
Minor bugfixes and cleanups.
https://github.com/stevegrubb/libcap-ng/releases/tag/v0.8.3

Current certificate chain on the server people.redhat.com is
still broken. Reported to both Redhat and the developer.

The tarballs on github are not release tarballs in the
"make dist" sense. Just a repo archive.
So they dont hash the same as the release archive.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20220404084653.4102053-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>