pipewire: disable v4l2 integration
PipeWire allows to emulate V4L2 devices by intercepting the device
ioctls and handling them in PipeWire. This is unrelated to the V4L2
backend, which allows to use V4L2 devices in PipeWire.

Disable the integration, because applications should use PipeWire
directly instead of using the V4L2 API.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20220421151942.342129-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>