host-autotools-autoconf: Version bump. 2.69 -> 2.71.
Tools are starting to prereq version 2.71.
So might aswell swallow the bitter pill and upgrade autoconf.

* License file changed. Nothing but whitespace, misspelling and URL fixes.
* Add a sed line for a standalone line in the autoconf binary.
It's a full path trailer.m4, which didn't match with the previous rules.
Add a separate one as it might get removed either way in the next version.
Looks like the autoconf team is planning a rewrite in the generation mechanism
using perl instead(?).

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20220513125653.955993-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>