ptxd_make_world_init: set invalid proxy value for download prevention to greppable value
In case a download is triggered in a non-get stage, e.g. in a code
signing provider missing the ptx/online-code-signing-provider macro,
a proxy error is printed in most cases.

Help the confused developers by giving them an invalid proxy value that
is self-explanatory and greppable.

In my case the proxy error message from the online code signing provider
was:

  proxyconnect tcp: dial tcp: lookup -: no such host

With this patch it is now:

   proxyconnect tcp: dial tcp: lookup PTXDIST-UNALLOWED-DOWNLOAD: no such host

Signed-off-by: Bastian Krause <bst@pengutronix.de>
Message-Id: <20220520153210.7158-1-bst@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>