barebox: Add option to install barebox generated dtbs.
Sometimes it's useful to be able to include the barebox-
generated dtb. An example would be the FIP image target
for TF-A and stm32mp1 targets.

Prefix the barebox-installed dtbs with "barebox-" as they can
have the same name as kernel dtbs for example.

This can partly go away if/when barebox decides to promote
dtbs to the barebox-flash-images file.
Although one would probably still want to rename the dtbs
to avoid a name collision.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20220609213328.3913523-2-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>