u-boot: Use host pkg-config to find host tool dependencies
Since U-Boot version 2015.07 U-Boot host tools might require
host-openssl, depending on options set in U-Boot Kconfig.

An option to depend on host openssl was added to ptxdist with commit
0c779a60e50c ("u-boot: Add option to depend on host-openssl") already.
This is not enough however, because 'sysroot-cross/bin/pkg-config' is
used instead of 'sysroot-host/bin/pkg-config' to discover libssl
dependencies.  This might not lead to build failures right away because
U-Boot has a fallback in its tools/Makefile masking the problem.  The
problem might also be masked if a libssl-dev is installed on the build
host (outside of ptxdist).

Using the same approach as barebox package template does since commit
1d7a4f8374d0 ("template-barebox-make: make sure the correct pkg-config
is used") lets ptxdist call the correct pkg-config.

Note: you can not backport this change to ptxdist lower than 2018.09.0,
see link below for detailed discussion.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Link: https://lore.ptxdist.org/ptxdist/10086649.DBdDmKp7kz@ada/T/#t
Message-Id: <20220803131114.6722-1-ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>