ptxdist: create and use diffs for 'ptxdist setup'
We have defaults in config/setup/ptxdistrc.default. But as soon as the user
runs 'ptxdist setup', a local config is created and used. This means that
changes to the defaults will be ignored because the local config still
contains the old values.

To avoid this, use the layer config diff mechanism to create a diff for the
ptxdistrc and use that in combination with ptxdistrc.default. This way only
explicitly changed variables are overwritten and defaults for others can be
updated.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>