git-ptx-patches: fix series parsing
If the line for the specified tag ends with the tag name (e.g. it does not
contain a ' --start-number ..' argument) then the next tag in the series
file is not detected.
As a result, all following commits are associated with this tag.

Improve the regex to fix this and rewrite the whole parsing to avoid such
errors in the future.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>