host-opkg-utils: Fix tarball format.
git.yoctoproject.org killed off their snapshot functionality yesterday.
Apparently too taxing to handle on the fly, thinking that nothing but
web crawlers was using snapshot. Not realizing that even the yocto poky
recipe was still using it.

After pointing it out, they've re-enabled the functionality.
But only for gz.
Their recipe will probably be rewritten to do git fetches instead.

So expect this to move to git only in the future.
But for now, just move it to .gz instead of .bz2.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20221006142216.3469201-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>