Revert "systemd: move dependencies to the toplevel option"
While this is technically, it causes some problems:
For !RUNTIME SYSTEMD_HWDB is disabled as it has no prompt. As a result, the
necessary build-time dependency SYSTEMD_HWDB -> HOST_SYSTEMD is lost.
It it is selected unconditionally then this creates cyclic dependencies.
So revert this and add it to the lint exception list.

This reverts commit 8b17816376e328401030d45cadc3a0fd86302f54.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>