modemmanager: version bump 1.18.12 -> 1.20.0
There is no longer an upstream-provided source tarball.
This is the last stable release with autoconf support
which is considered deprecated and will be removed in
future releases in favour of meson, so do the change now.

dbus-1 is now mandatory dependency.

This release comes with lots of new features, see announcement:
https://lists.freedesktop.org/archives/modemmanager-devel/2022-October/009481.html

Signed-off-by: Ladislav Michl <ladis@linux-mips.org>
Message-Id: <Y1xN4Fyv93f6PNOK@lenoch>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>