ptxd_make_xpkg_pkg: avoid calling mkdir if possible
Executing forking and executing mkdir often takes a lot of time. And in
most cases, the directory already exists. So only execute mkdir if the
directory does not exist yet.
And only create the lockfile directory once per package.
For packages with many files, this can save a lot of time in targetinstall.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>