install_copy_toolchain: fix with crosstool-ng
With crosstool-ng, for some libraries (e.g. libgcc_s.so.1), the result of
-print-file-name= is not in the sysroot but a separate directory with
individual symlinks to the sysroot.
In this case, the file name (without directory) of the link itself and the
link target are identical.
This confuses the logic that tries to generate symlinks for the BSP. As a
result, a symlink that points to itself is created.

Since 11038a892ada9af914c4e0a0de2951fa105b14af this symlink is not removed
before it is overwritten by the actuall file and overwriting may fail.

So make sure that the bogus symlink is not created at all.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>