libcamera: install IPA proxy workers unconditionally
If an IPA module is unsigned or the signature is invalid, the IPA is
isolated and runs in a separate process.
These so-called proxy workers must be installed for each activated
pipeline, even if the corresponding IPA module is not selected. Otherwise,
an out-of-tree IPA module (these are unsigned) cannot be used.
Since there is not an IPA module or associated proxy for every pipeline
(e.g. simple, uvcvideo), an additional list is required.

Signed-off-by: Matthias Fend <matthias.fend@emfend.at>
Message-Id: <20221116094045.202386-1-matthias.fend@emfend.at>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>