mariadb: fix building with cmake 3.25.0
The check if symbol versioning works creates an extra file. With cmake
3.25.0 each test is executed in its own temporary directory. As a result
the file is created in the wrong directory and the check fails.

We don't need symbol versioning in ptxdist so just disable it. In newer
mariadb versions, this code is gone anyways and this workaround can be
removed.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>