ptxd_make_get_svn: Sort members of resulting archive by name
Without this option the order of the files in the created archive
depends on things like the used filesystem. To make the order (and so
the resulting md5sum) reproducible, use --sort=name.

This option is supported in GNU tar since version 1.28 from 2014, which
is available since Debian 7 (aka wheezy).

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Message-Id: <20221205172824.3468159-1-u.kleine-koenig@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>