wrapper: fix building package from local_src/ symlinks
When the sources of a package are from a symlink in local_src/ then the
absolute path for those sources may be outside the BSP.
In this case, it can happen that a package adds such an absolute include
path to the search list. This path is filtered out incorrectly.
This happens for example with the linux kernel.

So allow search paths within the package sources with all symlinks
resolved.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>