ptxd_make_world_package_info: fix for packages like hostapd
Not all packages that define <PKG>_CONFIG use the config diff mechanism.
In this case the config may be located in ptxdist and
ptxd_make_kconfig_init fails because the config is not found within the
BSP.
So check for this failure and don't try to determine base/ref configs.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>