systemd: make default-net-naming-scheme configurable
The default-net-naming-scheme parameter allows to preserve the naming
scheme of a previous systemd version on a newer version. It may be
necessary to preserve a naming scheme if other system components rely on
certain network interface names generated by some naming scheme.

The default is "latest" to use the naming scheme of the currently used
systemd version.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20230110153021.1839663-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>