pulseaudio: fix targetinstall without daemon
The entire src/daemon subdirectory is disabled. This disables the build
of the pulseaudio executable and the generation of the daemon.conf,
system.pa, and default.pa configuration files. As the daemon is not
build, the D-BUS bus configuration and the systemd.service won't be
generated, too.

In src/utils, the pacmd and pasuspender are explicitly disabled.
pasuspender was never installed by the ptxdist rule.

The src/pulsecore subdirectory is disabled. Therefore, libpulsecore is
not built.

The src/modules subdirectory is disabled. Therefore, no modules are
built. As this disables the alsa-mixer module, there are no
configuration files for the alsa-mixer and, thus, the
usr/share/pulseaudio directory is empty and not created.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20230111094944.3388707-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>