ptxdist: make sure that the BSP git dir is ignored
The toplevel BSP directory is a indirect parent of all package source
directories. This means that all git operations within the package sources
will succeed without error. This can be annoying during development (e.g.
"git grep ..." never returns anything). It is also a problem with packages
that default to git magic (usually something with "git describe") to
determine the version.

The empty (invalid) .git file will trigger an error when git searches for
the git directory, so a git directory in the BSP is never found.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>