systemd: add option to help with systemd-timesyncd on nfsroot
On nfsroot with a read-only rootfs, systemd-timesyncd fails to start
because /var/lib/systemd/timesync is not owned by systemd-timesync.

In most cases with a read-only rootfs, /var/lib/systemd/timesync is not
writable anyways, so it doesn't really matter if it exists with the right
permissions anyways.
So add an option for this that clears StateDirectory= and allows
systemd-timesyncd to start.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>