image-hdimg: support using barebox with GPT
Currently, choosing IMAGE_HDIMG_GPT together with IMAGE_HDIMG_BAREBOX
leads to an error in genimage because the full GPT array is larger than
the hole in the barebox image. However, the hole in the barebox images
is large enough to fit the protective MBR and the primary GPT header.
The rest of the GPT array needs to moved after the barebox image, so set
the gpt-location setting to the size of the barebox image.

For the other bootloaders, nothing needs to change, so set the
gpt-location to its default value of 1024.

For MBR, this config setting is ignored.

Link: https://www.barebox.org/doc/latest/boards/imx.html#using-gpt-on-i-mx
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20230316165132.423488-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>