python3: fix include paths in sysconfigdata
The paths here are without sysroot, so they point to arbitrary paths in
the host system. It still works sometimes because fallbacks are used
when the path does not exist. But when e.g. /usr/include/python3.11
exists then it is used and later dropped by the wrapper. As a result,
Python headers are not found.
Fix this by adding the sysroot to all include paths.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>