ptxd_make_world_extract_cargo: handle workspaces and git URLs
workspace packages cannot be added to the local source directory
directly. The individual packages must be added instead.

Git URLs need an extra section in the config to redirect to the local
source directory. And the checksum in .cargo-checksum.json must not be
set. Otherwise cargo wants to add it to the Cargo.lock file and refuses
to build with --frozen.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>