systemd: match end0 interface in default network rule
Since systemd v252, the network interface name is generated based on the
device tree. Thus, the default network interface may no longer be named
eth0.

Make sure that DHCP is used on the renamed interface, too.

Rename the file to default.network to avoid the interface name in the
file name.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20230427121133.236696-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>