qemu / 9p: store metadata for / in correct file
Since qemu commit 81ffbf5ab1458e357a761f1272105a55829b351e the metadata
for / is not stored in platform-<nnn>/.virtfs_metadata/root but instead
in platform-<nnn>/root/.virtfs_metadata_root. For nfsroot, we also leave
the file in the current position.

Without this patch, / belongs to the user+group of the ptxdist user on
9p, not to root+root.

Inspired-by: Michael Olbrich <m.olbrich@pengutronix.de>
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>
Message-Id: <20230428080219.2971043-1-r.schwebel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>