doc: css: allow word-wrapping in grid table cells
By default, spinx_rtd_theme sets the CSS attribute "white-space: nowrap"
on table cells, which has the effect, that table contents never
word-wrap and the table gets scrollable. This is unfortunate if table
cells contain a large amount of text.

Issue #1246 of sphinx_rtd_theme addresses this situation, but the
general fix is not that easy. Until this issue is fixed upstream,
introduce a custom CSS class named "word-wrap" that reverts the
behaviour of table cells back to do word wrapping when the content of
the cell is wider than the available width.

With this change, the following example should produce a table as wide
as the content area with word-wrapped text in the second row:

  .. table::
     :class: word-wrap

     +--------------------------------+-----------------------------+
     | Column 1                       | Column 2                    |
     +================================+=============================+
     | Lorem ipsum dolor sit amet,    | Duis aute irure dolor in    |
     | adipisicing elit, sed do eius  | voluptate velit esse cillum |
     | incididunt ut labore et dolor  | fugiat nulla pariatur.      |
     | aliqua. Ut enim ad minim veni  | Excepteur sint occaecat     |
     | nostrud exercitation ullamco   | proident, sunt in culpa     |
     | nisi ut aliquip ex ea commodi. | deserunt mollit anim id.    |
     +--------------------------------+-----------------------------+
     | Third Row                      | Third Row continued         |
     +--------------------------------+-----------------------------+

Link: https://github.com/readthedocs/sphinx_rtd_theme/issues/1246
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20230502075505.316900-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>