ptxd_make_world_extract: fix compile_commands.json with lndir URL
We create a compile_commands.json in the source dir from the
compile_commands.json in the target dir. If the package is
extracted again, then the compile_commands.json in the target dir is a
symlink to the one in the source dir. This breaks the whole mechanism,
so remove the symlink while "extracting" the package.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>